Add: Gamesディレクトリ内の一部のファイルをMyPagesに移動させた
