+++ frontend/src/components/MyPages/FastAnalysisBox.tsx
