+++ frontend/src/components/MyPages/GameClearCountBox.tsx
