+++ frontend/src/components/MyPages/StudyHeatMap.tsx
