+++ frontend/src/containers/MyPages.tsx
@@ -1,54 +1,61 @@
-import React, { Fragment, useContext, useEffect, useLayoutEffect, useState, useMemo } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
+import React, {
+  Fragment,
+  useContext,
+  useEffect,
+  useLayoutEffect,
+  useState,
+  useMemo,
+} from "react";
+import { useLocation, useNavigate } from "react-router-dom";
+import styled from "styled-components";
 
 // Presentational Components
-import { Header } from '../components/Headers/Header';
-import { SessionFlashMessage } from '../components/FlashMessages/SessionFlashMessage';
-import { UserStatus } from '../components/Games/UserStatus';
-import { StudyHeatMap } from '../components/Games/StudyHeatMap';
-import { DescriptionWrapper } from '../components/shared_style';
-import { GameContent } from '../components/GameContents/GameContent';
-import { TitleCard } from '../components/Games/TitleCard';
-import { Footer } from '../components/Footers/Footer';
-import { ReleaseConditionDialog } from '../components/Dialogs/ReleaseConditionDialog'
-import { CircularMask } from '../components/loads/CircularMask';
-import { GameClearCountBox } from '../components/Games/GameClearCountBox';
-import { TimeAnalysisBox } from '../components/Games/TimeAnalysisBox';
-import { FastAnalysisBox } from '../components/Games/FastAnalysisBox';
-import { SecondSelectBox } from '../components/Games/SecondSelectBox';
-import { TitleFlashMessage } from '../components/FlashMessages/TitleFlashMessage';
+import { Header } from "../components/Headers/Header";
+import { SessionFlashMessage } from "../components/FlashMessages/SessionFlashMessage";
+import { UserStatus } from "../components/MyPages/UserStatus";
+import { StudyHeatMap } from "../components/MyPages/StudyHeatMap";
+import { DescriptionWrapper } from "../components/shared_style";
+import { GameContent } from "../components/GameContents/GameContent";
+import { TitleCard } from "../components/MyPages/TitleCard";
+import { Footer } from "../components/Footers/Footer";
+import { ReleaseConditionDialog } from "../components/Dialogs/ReleaseConditionDialog";
+import { CircularMask } from "../components/loads/CircularMask";
+import { GameClearCountBox } from "../components/MyPages/GameClearCountBox";
+import { TimeAnalysisBox } from "../components/MyPages/TimeAnalysisBox";
+import { FastAnalysisBox } from "../components/MyPages/FastAnalysisBox";
+import { SecondSelectBox } from "../components/MyPages/SecondSelectBox";
+import { TitleFlashMessage } from "../components/FlashMessages/TitleFlashMessage";
 
 // Images
-import ElementaryGameContentImage from '../images/elementary_game_content.jpg';
-import IntermediateGameContentImage from '../images/intermediate_game_content.jpg';
-import AdvancedGameContentImage from '../images/advanced_game_content.jpg';
+import ElementaryGameContentImage from "../images/elementary_game_content.jpg";
+import IntermediateGameContentImage from "../images/intermediate_game_content.jpg";
+import AdvancedGameContentImage from "../images/advanced_game_content.jpg";
 
 // Contextオブジェクト
 import { UserContext } from "../context/UserProvider";
 
 // ログイン状態を確認するAPIコール関数
-import { checkLoginStatus } from '../apis/checkLoginStatus';
+import { checkLoginStatus } from "../apis/checkLoginStatus";
 
 // マイページの情報を取得するAPIコール関数
-import { getMyPageInfo } from '../apis/mypage';
+import { getMyPageInfo } from "../apis/mypage";
 
 // HTTP_STATUS_CODE
-import { HTTP_STATUS_CODE } from '../constants';
+import { HTTP_STATUS_CODE } from "../constants";
 
 // REQUEST_STATE
-import { REQUEST_STATE } from '../constants';
+import { REQUEST_STATE } from "../constants";
 
 // Colors
-import { COLORS } from '../style_constants';
+import { COLORS } from "../style_constants";
 
 // 今月の月を取得する関数
-import { getMonthOfTheMonth } from '../functions/getMonthOfTheMonth';
+import { getMonthOfTheMonth } from "../functions/getMonthOfTheMonth";
 
-import { WIDTH } from '../style_constants';
+import { WIDTH } from "../style_constants";
 
 // myPageStateの型
-import { MyPageState } from '../types/containers/myPages';
+import { MyPageState } from "../types/containers/myPages";
 
 // メインのラッパー
 const MainWrapper = styled.div`
@@ -102,7 +109,7 @@ const SecondWrapper = styled.div`
   margin: 0 auto;
   @media (max-width: ${WIDTH.DEV_TOOL}) {
     width: 92%;
-  };
+  } ;
 `;
 
 // サードラッパー
@@ -127,7 +134,7 @@ const StudyHeatMapSentenceWrapper = styled(DescriptionWrapper)`
   padding-left: 12.5%;
   @media (max-width: ${WIDTH.DEV_TOOL}) {
     font-size: 1.3em;
-  };
+  } ;
 `;
 
 // クエスト一覧というセンテンスのラッパー
@@ -138,7 +145,7 @@ const QuestSentenceWrapper = styled(DescriptionWrapper)`
   text-align: center;
   @media (max-width: ${WIDTH.DEV_TOOL}) {
     font-size: 1.35em;
-  };
+  } ;
 `;
 
 // マイページのゲームコンテンツのラッパー
@@ -150,7 +157,7 @@ const MyPageGameContentsWrapper = styled.div`
   flex-wrap: wrap;
   @media (max-width: ${WIDTH.DEV_TOOL}) {
     justify-content: center;
-  };
+  } ;
 `;
 
 // 称号一覧というセンテンスのラッパー
@@ -174,7 +181,6 @@ const TitleListWrapper = styled.div`
 `;
 
 export const MyPages = (): JSX.Element => {
-
   // useContext
   // requestUserStateには、requestState, userState, errorsが格納されている
   // userStateにはsessionとuserが格納されている
@@ -183,10 +189,10 @@ export const MyPages = (): JSX.Element => {
       requestState,
       sessionState,
       userState: { user },
-      battleAudioState
+      battleAudioState,
     },
     dispatch,
-    requestUserActionTyps
+    requestUserActionTyps,
   } = useContext(UserContext);
 
   // 今月の月を計算する関数
@@ -215,13 +221,13 @@ export const MyPages = (): JSX.Element => {
     totalTimePerDifficulty: {
       advanced: 0,
       elementary: 0,
-      intermediate: 0
+      intermediate: 0,
     },
     gameClearCountPerDifficulty: {
       advanced: 0,
       elementary: 0,
-      intermediate: 0
-    }
+      intermediate: 0,
+    },
   };
 
   // MyPageの状態を管理するstate
@@ -235,34 +241,35 @@ export const MyPages = (): JSX.Element => {
 
   // ブラウザをリロードしてもログイン状態を維持するためのuseEffect
   useEffect(() => {
-    if(sessionState === false){
+    if (sessionState === false) {
       dispatch({ type: requestUserActionTyps.REQUEST });
-      checkLoginStatus().then((data) => {
-        dispatch({
-          type: requestUserActionTyps.REQUEST_SUCCESS,
-          payload: {
-            session: data.session,
-            user: data.user,
+      checkLoginStatus()
+        .then((data) => {
+          dispatch({
+            type: requestUserActionTyps.REQUEST_SUCCESS,
+            payload: {
+              session: data.session,
+              user: data.user,
+            },
+          });
+          if (!data.session && location.key === "default") {
+            navigate("/", {
+              state: { display: true, success: "ログインしてください。" },
+            });
+          }
+        })
+        .catch((e) => {
+          if (e.response.status === HTTP_STATUS_CODE.NOT_FOUND) {
+            dispatch({
+              type: requestUserActionTyps.REQUEST_FAILURE,
+              payload: {
+                errors: e.response.data.errors,
+              },
+            });
+          } else {
+            throw e;
           }
         });
-        if(!data.session && location.key === 'default') {
-          navigate(
-            '/',
-            { state: { display: true, success: "ログインしてください。"}}
-          )
-        }
-      }).catch((e) => {
-        if(e.response.status === HTTP_STATUS_CODE.NOT_FOUND){
-          dispatch({
-            type: requestUserActionTyps.REQUEST_FAILURE,
-            payload: {
-              errors: e.response.data.errors
-            }
-          });
-        } else {
-          throw e;
-        }
-      })
     }
   }, [
     dispatch,
@@ -271,7 +278,7 @@ export const MyPages = (): JSX.Element => {
     requestUserActionTyps.REQUEST_SUCCESS,
     requestUserActionTyps.REQUEST_FAILURE,
     navigate,
-    location.key
+    location.key,
   ]);
 
   // マイページの情報を取得するuseLayoutEffect
@@ -282,171 +289,158 @@ export const MyPages = (): JSX.Element => {
   // getPageInfoのおかげで、マイページのデータを取得したら、
   // このuseLayoutEffectは機能しないように設定されている
   useLayoutEffect(() => {
-    if(!myPageState.getPageInfo && sessionState && Object.keys(user).length){
-      getMyPageInfo(user.id as number).then((data) => {
-        setMyPageState((prev) => ({
-          ...prev,
-          gameFrequenciesPerDay: data.game_frequencies_per_day,
-          totalTimePerDifficulty: data.total_time_per_difficulty,
-          gameClearCountPerDifficulty: data.game_clear_count_per_difficulty,
-          fastTimePerDifficulty: data.fast_time_per_difficulty,
-          ownedTitles: data.owned_titles,
-          selectedTotalTime: data.total_time_per_difficulty.elementary,
-          selectedGameClearCount: data.game_clear_count_per_difficulty.elementary,
-          selectedFastTime: data.fast_time_per_difficulty.elementary,
-          getPageInfo: true
-        }));
-      }).catch((e) => {
-        if(e.response.status === HTTP_STATUS_CODE.NOT_FOUND){
+    if (!myPageState.getPageInfo && sessionState && Object.keys(user).length) {
+      getMyPageInfo(user.id as number)
+        .then((data) => {
           setMyPageState((prev) => ({
             ...prev,
+            gameFrequenciesPerDay: data.game_frequencies_per_day,
+            totalTimePerDifficulty: data.total_time_per_difficulty,
+            gameClearCountPerDifficulty: data.game_clear_count_per_difficulty,
+            fastTimePerDifficulty: data.fast_time_per_difficulty,
+            ownedTitles: data.owned_titles,
+            selectedTotalTime: data.total_time_per_difficulty.elementary,
+            selectedGameClearCount:
+              data.game_clear_count_per_difficulty.elementary,
+            selectedFastTime: data.fast_time_per_difficulty.elementary,
+            getPageInfo: true,
           }));
-        } else {
-          throw e;
-        }
-      });
+        })
+        .catch((e) => {
+          if (e.response.status === HTTP_STATUS_CODE.NOT_FOUND) {
+            setMyPageState((prev) => ({
+              ...prev,
+            }));
+          } else {
+            throw e;
+          }
+        });
     }
-  }, [
-    user,
-    sessionState,
-    thisMonth,
-    myPageState.getPageInfo
-  ]);
+  }, [user, sessionState, thisMonth, myPageState.getPageInfo]);
 
   // ゲーム中のユーザーがトップページに戻ったときに
   // 音を消すuseEffect
   useEffect(() => {
-    if(battleAudioState.play) {
+    if (battleAudioState.play) {
       battleAudioState.audio.pause();
       battleAudioState.audio.currentTime = 0;
     }
-  },[
-    battleAudioState.play,
-    battleAudioState.audio
-  ])
+  }, [battleAudioState.play, battleAudioState.audio]);
 
   return (
     <>
-      {
-        requestState === REQUEST_STATE.LOADING ?
-          <CircularMask />
-        :
-          <>
-            <Header />
-            <MainWrapper>
-              <SessionFlashMessage
-                location={location}
+      {requestState === REQUEST_STATE.LOADING ? (
+        <CircularMask />
+      ) : (
+        <>
+          <Header />
+          <MainWrapper>
+            <SessionFlashMessage location={location} />
+            <TitleFlashMessage
+              display={myPageState.display}
+              message={myPageState.message}
+              setMyPageState={setMyPageState}
+            />
+            <MainFirstWrapper>
+              <UserStatus
+                name={user.name as string}
+                rank={user.rank}
+                activeTitle={user.active_title}
+                temporaryExperience={user.temporary_experience}
+                totalExperience={user.total_experience}
+                maximumExperiencePerRank={user.maximum_experience_per_rank}
+                image={user.image}
               />
-              <TitleFlashMessage
-                display={myPageState.display}
-                message={myPageState.message}
-                setMyPageState={setMyPageState}
-              />
-              <MainFirstWrapper>
-                <UserStatus
-                  name={user.name as string}
-                  rank={user.rank}
-                  activeTitle={user.active_title}
-                  temporaryExperience={user.temporary_experience}
-                  totalExperience={user.total_experience}
-                  maximumExperiencePerRank={user.maximum_experience_per_rank}
-                  image={user.image}
+              <StudyHeatMapWrapper>
+                <StudyHeatMapSentenceWrapper>
+                  学習カレンダー
+                </StudyHeatMapSentenceWrapper>
+                <StudyHeatMap
+                  gameFrequenciesPerDay={myPageState.gameFrequenciesPerDay}
                 />
-                <StudyHeatMapWrapper>
-                  <StudyHeatMapSentenceWrapper>
-                    学習カレンダー
-                  </StudyHeatMapSentenceWrapper>
-                  <StudyHeatMap
-                    gameFrequenciesPerDay={myPageState.gameFrequenciesPerDay}
-                  />
-                </StudyHeatMapWrapper>
-              </MainFirstWrapper>
-              <MainSecondWrapper>
-                <MainSecondSelectWrapper>
-                  <SecondSelectBox
-                    difficultyMonthTitle={myPageState.difficultyMonthTitle}
-                    prevDifficultyMonthTitle={myPageState.prevDifficultyMonthTitle}
-                    nextDifficultyMonthTitle={myPageState.nextDifficultyMonthTitle}
-                    setMyPageState={setMyPageState}
-                    thisMonth={thisMonth}
-                  />
-                </MainSecondSelectWrapper>
-                <MainSecondGraphWrapper>
-                  <TimeAnalysisBox
-                    time={myPageState.selectedTotalTime}
-                  />
-                  <GameClearCountBox
-                    count={myPageState.selectedGameClearCount}
-                  />
-                  <FastAnalysisBox
-                    minutes={myPageState.selectedFastTime}
-                  />
-                </MainSecondGraphWrapper>
-              </MainSecondWrapper>
-              <SecondWrapper>
-                <QuestSentenceWrapper>
-                  クエスト一覧
-                </QuestSentenceWrapper>
-                <MyPageGameContentsWrapper>
-                  <GameContent
-                    difficulty='elementary'
-                    image={ElementaryGameContentImage}
-                  />
-                  <GameContent
-                    difficulty='intermediate'
-                    image={IntermediateGameContentImage}
-                  />
-                  <GameContent
-                    difficulty='advanced'
-                    image={AdvancedGameContentImage}
-                  />
-                </MyPageGameContentsWrapper>
-              </SecondWrapper>
-              <ThirdWrapper>
-                <TitleListSentenceWrapper>
-                  称号一覧
-                </TitleListSentenceWrapper>
-                <TitleListWrapper>
-                  {
-                    myPageState.ownedTitles.map((obj, index) => (
-                      <Fragment key={index}>
-                        <TitleCard
-                          name={obj.name}
-                          releaseDate={obj.release_date}
-                          onClick={() => setMyPageState((prev) => ({
-                            ...prev,
-                            isOpenDialog: true,
-                            name: obj.name,
-                            releaseDate: obj.release_date,
-                            releaseCondition: obj.release_condition
-                          }))}
-                        />
-                      </Fragment>
-                    ))
+              </StudyHeatMapWrapper>
+            </MainFirstWrapper>
+            <MainSecondWrapper>
+              <MainSecondSelectWrapper>
+                <SecondSelectBox
+                  difficultyMonthTitle={myPageState.difficultyMonthTitle}
+                  prevDifficultyMonthTitle={
+                    myPageState.prevDifficultyMonthTitle
                   }
-                </TitleListWrapper>
-              </ThirdWrapper>
-            </MainWrapper>
-            <Footer />
-            {
-              myPageState.isOpenDialog &&
-                <ReleaseConditionDialog
-                  isOpen={myPageState.isOpenDialog}
-                  onClose={() => setMyPageState((prev) => ({
-                    ...prev,
-                    isOpenDialog: false,
-                    name: "",
-                    releaseDate: "",
-                    releaseCondition: ""
-                  }))}
-                  name={myPageState.name}
-                  releaseDate={myPageState.releaseDate}
-                  releaseCondition={myPageState.releaseCondition}
+                  nextDifficultyMonthTitle={
+                    myPageState.nextDifficultyMonthTitle
+                  }
                   setMyPageState={setMyPageState}
+                  thisMonth={thisMonth}
                 />
-            }
-          </>
-      }
+              </MainSecondSelectWrapper>
+              <MainSecondGraphWrapper>
+                <TimeAnalysisBox time={myPageState.selectedTotalTime} />
+                <GameClearCountBox count={myPageState.selectedGameClearCount} />
+                <FastAnalysisBox minutes={myPageState.selectedFastTime} />
+              </MainSecondGraphWrapper>
+            </MainSecondWrapper>
+            <SecondWrapper>
+              <QuestSentenceWrapper>クエスト一覧</QuestSentenceWrapper>
+              <MyPageGameContentsWrapper>
+                <GameContent
+                  difficulty="elementary"
+                  image={ElementaryGameContentImage}
+                />
+                <GameContent
+                  difficulty="intermediate"
+                  image={IntermediateGameContentImage}
+                />
+                <GameContent
+                  difficulty="advanced"
+                  image={AdvancedGameContentImage}
+                />
+              </MyPageGameContentsWrapper>
+            </SecondWrapper>
+            <ThirdWrapper>
+              <TitleListSentenceWrapper>称号一覧</TitleListSentenceWrapper>
+              <TitleListWrapper>
+                {myPageState.ownedTitles.map((obj, index) => (
+                  <Fragment key={index}>
+                    <TitleCard
+                      name={obj.name}
+                      releaseDate={obj.release_date}
+                      onClick={() =>
+                        setMyPageState((prev) => ({
+                          ...prev,
+                          isOpenDialog: true,
+                          name: obj.name,
+                          releaseDate: obj.release_date,
+                          releaseCondition: obj.release_condition,
+                        }))
+                      }
+                    />
+                  </Fragment>
+                ))}
+              </TitleListWrapper>
+            </ThirdWrapper>
+          </MainWrapper>
+          <Footer />
+          {myPageState.isOpenDialog && (
+            <ReleaseConditionDialog
+              isOpen={myPageState.isOpenDialog}
+              onClose={() =>
+                setMyPageState((prev) => ({
+                  ...prev,
+                  isOpenDialog: false,
+                  name: "",
+                  releaseDate: "",
+                  releaseCondition: "",
+                }))
+              }
+              name={myPageState.name}
+              releaseDate={myPageState.releaseDate}
+              releaseCondition={myPageState.releaseCondition}
+              setMyPageState={setMyPageState}
+            />
+          )}
+        </>
+      )}
     </>
   );
 };
